requirements.txt
python-telegram-bot==13.14
boto3==1.25.4